Python/requirements.txt
discord.py>=2.3.0
python-telegram-bot>=20.3
python-dotenv>=1.0.0